chore: organize codebase features into modules / crates
