+++ .env.example
@@ -1,7 +1,8 @@
-# Server Configuration
-PORT=8080
-HOST=127.0.0.1
+# Tycho API settings
+TYCHO_URL=tycho-beta.propellerheads.xyz
+TYCHO_API_KEY=your_api_key_here
+TVL_THRESHOLD=1000
 
-# Add other environment variables here
-TYCHO_API_KEY=APIKEY
-RPC_URL=RPC_URL
+# Server settings
+PORT=3000
+HOST=127.0.0.1
